+++ src/bayestar_sky_map.h
@@ -97,7 +97,7 @@ bayestar_pixel *bayestar_sky_map_toa_phoa_snr(
     double gmst,                    /* GMST (rad) */
     unsigned int nifos,             /* Number of detectors */
     unsigned long nsamples,         /* Lengths of SNR series */
-    double sample_rate,             /* Sample rate in seconds */
+    float sample_rate,              /* Sample rate in seconds */
     const double *epochs,           /* Timestamps of SNR time series */
     const float complex **snrs,     /* Complex SNR series */
     const float (**responses)[3],   /* Detector responses */
@@ -134,11 +134,11 @@ double bayestar_log_posterior_toa_phoa_snr(
  * This is more of an internal function, but it's *really* important that
  * it agrees with LAL conventions, so we expose it in the interface in order
  * to be to validate it in Python against the LALSimulation SWIG bindings. */
-double complex bayestar_signal_amplitude_model(
-    double complex F,               /* Complex antenna factor */
-    double complex exp_i_twopsi,    /* e^(i*2*psi), for polarization angle psi */
-    double u,                       /* cos(inclination) */
-    double u2                       /* cos^2(inclination */
+float complex bayestar_signal_amplitude_model(
+    float complex F,               /* Complex antenna factor */
+    float complex exp_i_twopsi,    /* e^(i*2*psi), for polarization angle psi */
+    float u,                       /* cos(inclination) */
+    float u2                       /* cos^2(inclination */
 );
 
 /* Unit test suite. Return EXIT_SUCCESS if tests passed,
